Factorize containerized react component
